Extract CardGrid and reuse it in companies listing
This brings the same performance improvement to the companies card grid
as previous commit did for municipalities.

It also cleans up a left-over distinction from when we had companies
grouped under industry.
